requirements.txt
alembic==1.5.6
aniso8601==9.0.1
appdirs==1.4.4
attrs==20.3.0
bcrypt==3.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dataclasses==0.8
flake8==3.8.4
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
idna==2.10
importlib-metadata==3.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
regex==2020.11.13
requests==2.25.1
# Editable Git install with no remote (Robots-management==0.1)
-e /Users/martinpavic/Desktop/GideonTask/src
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
zipp==3.4.1
